corrected api server volume issue
When trying to run the server, I was running into errors when trying to start up the API. I traced the issue down to a line in server/docker-compose.yml. On line 19, I changed the volume to select "index" rather than "app", as "app" does not exist, and the errors were resolved. The API now boots up successfully.